feat(app-5nu): Т30 — шов предсказания движения и оружия
PlayerPrediction.Step — публичный шов Simulation (спека §3.9): sanitize →
AimPoint → движение → оружие без спавна. Гейт краевых запросов не дублируется
(живёт внутри PlayerMovementSystem.Update, Т10), MovementResult отбрасывается —
события и статистика серверные.

WeaponSystem: одно ядро Advance с опциональным стоком выстрелов (worldOrNull) —
null пропускает ровно снаряд, розыгрыш конуса и ShotsFired, всё остальное идёт
тем же порядком float-операций. Предикат вынесен в единственный дом CanFire;
Update сменил int index на byte ownerIndex.

PredictionParityTests (11 тестов): побитовое сравнение всех полей PlayerState
рефлексией после каждого тика, вход через InputCodec.Encode/Decode (Р34).
Сценарии: бег, рикошет дэша о стену, гашение слайда о стену, связка дэш↔слайд
под зажатой гашеткой, отказ по стамине, спам краевых запросов с обоими
счётчиками гейта, очередь на точной границе кулдауна и отпускание→перезажатие,
враждебный ввод через кодек и по сырому пути.

Спайк Т3 удалён полностью: Networking/Spike/**, NetSpike.unity, SpikePlayer.prefab,
SpikeSceneBootstrap.cs, PlayerMovementSpikeSeam, CompareSpikeReplicateData;
ссылки Ring.Networking/FishNet.Runtime убраны из Editor.asmdef.
MathCodegenSupport.cs оставлен (постоянный обход кодогена, Р110).

DefaultPrefabObjects.asset вычищен вручную: FishNet 4.7.2 GenerateFull гейтит
SetDirty/SaveAssets условием newCount > 0, поэтому опустевший реестр никогда не
пишется на диск — генератор физически не может убрать последний guid.

Прогон: 563/563, EXIT=0 (было 552, N=11); оба golden Passed и не сдвинулись
(0x5BD8AC0DE1D0C454 :503, 0x136FA6114112E44F :575), DeterminismTests.cs не тронут.
Мутационная волна — 8 из 8 уронены.

+++ client/Assets/Scripts/Simulation/Core/SimulationWorld.cs
@@ -207,7 +207,7 @@ namespace Ring.Simulation.Core
             for (int i = 0; i < _players.Length; i++)
             {
                 if (_players[i].Alive)
-                    WeaponSystem.Update(this, ref _players[i], in _sanitizedInputs[i], i);
+                    WeaponSystem.Update(this, ref _players[i], in _sanitizedInputs[i], (byte)i);
             }
             // Canonical tick order (spec Interfaces, Task 16/19/20): movement →
             // weapon → mobs (Phase 6) → mob separation → projectiles → (waves,
